+++ editor/src/lib.rs
@@ -14,4 +14,6 @@ pub mod native;
 #[cfg(not(target_arch = "wasm32"))]
 pub mod preferences;
 #[cfg(not(target_arch = "wasm32"))]
+pub mod project;
+#[cfg(not(target_arch = "wasm32"))]
 pub mod scene_file;